download_from_repo: Handle wanted pkg version being different
If we want pkg-1.2 and remote has pkg-1.1 then we can end up
bootstrapping pkg-1.1 and not keeping it around to use for
further internal needs. By specifying ports-mgmt/pkg we ensure
we get whatever version the remote has as our internal one.
A build for upgrade will still happen if needed.